i have some questions about forecasting_runner.py
Hey, bro, I don't quite understand the error correction of the input and output of the prediction part in the code. Why should the average value be added and subtracted in the input correction, and why should the Rho multiplied by the last column of X be added in the output correction? These things don't seem to be reflected in the paper. Can you help me solve them?My email is 1169844310@qq.com